summarizer/requirements.txt
aiohttp==3.5.4
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.78
botocore==1.12.78
bs4==0.0.1
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
docutils==0.14
gensim==3.6.0
idna==2.8
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.3.0
mccabe==0.6.1
multidict==4.5.2
netifaces==0.10.6
networkx==2.2
nltk==3.4
numpy==1.15.4
pandas==0.23.4
pylint==2.2.2
python-dateutil==2.7.5
python-engineio==3.2.3
python-socketio==3.1.1
pytz==2018.9
requests==2.21.0
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.7.1
soupsieve==1.7
urllib3==1.24.1
wrapt==1.11.0
yarl==1.3.0
